Count a run's attempts, because the object count never could
A finished report can carry more outcomes than the file has rows, from
two causes with opposite right answers: a stale dry-run skip whose row
then committed as a create (drop the skip), and a resume double-counting
after a checkpoint that did not persist (the surplus is in
created/updated, and taking it out of skipped erases a genuine refusal).

The comment said the attempt count tells them apart and the code read
`len(Report.Objects) == 1`. Attempts fold by object CLASS, so that
counts classes and never walks — and a CSV import is always exactly one
class, so the guard was true on every CSV report. Every resumed import
had its surplus taken out of `skipped`, dropping rows the report exists
to name.

Report carries Attempts, folded where the attempts are. Zero and one
both mean once, so a report written before this reads as the single
attempt it was.

engine.go passed the 500-line cap on the way, and the split it wanted
was the report: the folding is the part with a rule in it, and the run
loop is the part that walks.
